requirements.txt
wheel
ez_setup
gunicorn
loguru
pydantic
flask-openapi3
Flask-JWT-Extended
black
flake8
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask
pytest
google-cloud-tasks
firebase_admin
google-auth
